perf: optimize codebase
refactor: reuse components
