--- app/bot/dynamicWidget.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { DynamicWidget } from "@dynamic-labs/sdk-react-core";
-
-export default function Bot() {
-  return <DynamicWidget />;
-}
