+++ app/bot/layout.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 export default function BotLayout({
   children,
 }: Readonly<{
