+++ package.json
@@ -58,6 +58,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.51.1",
     "react-hot-toast": "^2.4.1",
+    "react-icons": "^5.2.1",
     "react-intersection-observer": "^9.10.3",
     "react-spring-bottom-sheet": "^3.4.1",
     "react-virtuoso": "^4.7.7",
